Add minimum amount for notification

+++ settlement_handler.go
@@ -72,7 +72,7 @@ func (sh *SettlementHandler) subscribeToInvoice(r_hash string, comment string) e
 				if err != nil {
 					log.Printf("Invalid amount %s.", invoice.Result.AmtPaidSat)
 				} else {
-					broadcastNotification(uint(amt), comment)
+					broadcastNotification(amt, comment)
 				}
 			}
 		}
